pretty-formats: enclose options in angle brackets
Enclose the 'options' placeholders in the %(describe) and %(trailers)
format specifiers in angle brackets to clarify that they are
placeholders rather than keywords.

Also remove the indentation from their descriptions, instead of
increasing it to account for the extra two angle bracket in the
headings. The indentation isn't required by asciidoc, it doesn't reflect
how the output text is formatted, and it's inconsistent with the
following bullet points that are at the same level in the output.

Signed-off-by: Andy Koppe <andy.koppe@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>